test: separate JQF fuzz tests into ValueFuzzTest and add missing emoji boundary
