+++ api/all/src/test/java/io/opentelemetry/api/logs/ValueTest.java
@@ -9,10 +9,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
-import edu.berkeley.cs.jqf.fuzz.Fuzz;
-import edu.berkeley.cs.jqf.fuzz.JQF;
-import edu.berkeley.cs.jqf.fuzz.junit.GuidedFuzzing;
-import edu.berkeley.cs.jqf.fuzz.random.NoGuidance;
 import io.opentelemetry.api.common.KeyValue;
 import io.opentelemetry.api.common.Value;
 import io.opentelemetry.api.common.ValueType;
@@ -30,8 +26,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.junit.runner.Result;
-import org.junit.runner.RunWith;
 
 class ValueTest {
 
@@ -213,32 +207,16 @@ class ValueTest {
 
   @ParameterizedTest
   @ValueSource(
-      strings = {"standard string", "special characters !@#$%^&*()", "multi\nline\tstring", ""})
+      strings = {
+        "standard string",
+        "special characters !@#$%^&*()",
+        "multi\nline\tstring",
+        "",
+        "emoji boundary test 🚀👩‍💻"
+      })
   void valueByteAsString(String inputStr) {
     String base64 = Value.of(inputStr.getBytes(StandardCharsets.UTF_8)).asString();
     byte[] decoded = Base64.getDecoder().decode(base64);
     assertThat(new String(decoded, StandardCharsets.UTF_8)).isEqualTo(inputStr);
-  }
-
-  @RunWith(JQF.class)
-  public static class FuzzTestCases {
-    @Fuzz
-    public void valueByteAsStringFuzz(String randomString) {
-      String base64Encoded = Value.of(randomString.getBytes(StandardCharsets.UTF_8)).asString();
-      byte[] decodedBytes = Base64.getDecoder().decode(base64Encoded);
-      assertThat(new String(decodedBytes, StandardCharsets.UTF_8)).isEqualTo(randomString);
-    }
-  }
-
-  @SuppressWarnings("SystemOut")
-  @Test
-  void valueByteAsStringFuzzing() {
-    Result result =
-        GuidedFuzzing.run(
-            FuzzTestCases.class,
-            "valueByteAsStringFuzz",
-            new NoGuidance(10000, System.out),
-            System.out);
-    assertThat(result.wasSuccessful()).isTrue();
   }
 }
